Added enemy and run button behaviour
Enemies can move left and right. Upon collision with the player, they trigger the beginning of a battle.
NOTE: There is a bug that makes it impossible to properly load in data after the 2nd battle. The fix is easy, but seeing as I need to move the code which generates the units anyway, it isn't worth fixing.