Added surprise attack behaviour
If player attacks an enemy, half of their strength is dealt as physical damage to each enemy. If the enemy sneaks up behind an enemy, half of the each enemy's strength is dealt as physical damage to each allied unit.

New SFX and animations have been added, but most are not being used yet.