Added enemy spawning. Modified battle scene enemies
In the platforming scene, enemies can now be spawned in at any coordinates with any movement speed.
Depending on what enemy was attacked in the platforming scene, we will load in that enemy in the battle scene and we will now generate its stats randomly based on the player's level.
Added new enemies: Masked Doctor, Cavern Monster and Serpent.